Update BookingSummary.vue
I added two rows at the bottom of the page to show both Refund and Privacy Policy.
I remove the refund policy link and set checkbox to false, this let user read and Accept refund policy and click to confirm they understand. The Accept refund policy landing page is not friendly.  
On the last row I have added the DPA statement with a checkbox and external link to full privacy policy.